Add error handling
