+++ server/app.js
@@ -15,7 +15,7 @@ const pollsRouter = require('./routes/polls/polls.route');
 // Friends routes
 const friendsRouter = require('./routes/users/friends/friends.route');
 // Friend List routes
-const friendListRouter = require('./routes/users/friendList/friendList.route');
+const friendListsRouter = require('./routes/friendLists/friendLists.route');
 
 const { json, urlencoded } = express;
 
@@ -54,7 +54,7 @@ app.use('/polls', pollsRouter);
 // Friends Routes
 app.use('/users/:userId/friends', friendsRouter);
 // Friend List Routes
-app.use('/users/:userId/friendList', friendListRouter);
+app.use('/friendLists', friendListsRouter);
 
 // Catch 404 and forward to error handler
 app.use(function (req, res, next) {
