+++ server/models/friendList.js
@@ -20,10 +20,4 @@ const FriendListSchema = new Schema({
   ],
 });
 
-FriendListSchema.methods.updateList = function (title, friendIds) {
-  this.title = title;
-  this.friendIds = friendIds.map((id) => mongoose.Types.ObjectId(id));
-  return this.save();
-};
-
 module.exports = FriendList = mongoose.model('friendList', FriendListSchema);
